some corrections

+++ src/antlr/v4/runtime/atn/LexerModeAction.d
@@ -103,9 +103,9 @@ class LexerModeAction : LexerAction
      * @safe
      * @nothrow
      */
-    public int hashCode() @safe nothrow
+    public size_t hashCode() @safe nothrow
     {
-	int hash = MurmurHash.initialize();
+	size_t hash = MurmurHash.initialize();
         hash = MurmurHash.update(hash, Utils.rank(getActionType));
         hash = MurmurHash.update(hash, mode);
         return MurmurHash.finish(hash, 2);
